+++ scripts/creatures/horrifying-visage.js
@@ -6,35 +6,32 @@
     Visage for the next 24 hours. The aging effect can be reversed with a greater restoration spell, but only within 24
     hours of it occurring.
  */
-const version = "10.0";
+const version = "11.0";
 const optionName = "Horrifying Visage";
 
 try {
-    const lastArg = args[args.length - 1];
-
-    if (lastArg.macroPass === "preambleComplete") {
-
-        // check target types
-        for (let target of lastArg.workflow.targets) {
+    if (args[0].macroPass === "preambleComplete") {
+        // check target types, remove the undead
+        for (let target of workflow.targets) {
             let creatureType = target.actor.system.details.type;
 
             // remove targets that are undead
             if ((creatureType != null) && (creatureType.value === "undead")) {
-                lastArg.workflow.targets.delete(target);
+                workflow.targets.delete(target);
             }
         }
 
-        game.user.updateTokenTargets(Array.from(lastArg.workflow.targets).map(t => t.id));
+        game.user.updateTokenTargets(Array.from(workflow.targets).map(t => t.id));
 
-    } else if (lastArg.macroPass === "postSave") {
-        let targets = lastArg.failedSaves;
+    } else if (args[0].macroPass === "postSave") {
+        let targets = workflow.failedSaves;
 
         if (targets && targets.length > 0) {
-            let saveDC = lastArg.itemData.system.save.dc;
+            let saveDC = itemData.system.save.dc;
 
             for (let target of targets) {
-                let targetIndex = lastArg.targets.indexOf(target);
-                let targetResult = lastArg.workflow.saveResults[targetIndex].total ?? 0;
+                let targetIndex = workflow.targets.indexOf(target);
+                let targetResult = workflow.workflow.saveResults[targetIndex].total ?? 0;
 
                 // compare the same result to the DC to determine if they should be aged
                 if (targetResult < (saveDC - 4)) {
@@ -49,7 +46,7 @@ try {
                 // Apply Frightened effect (overtime)
                 const effectData = {
                     label: `${optionName}`,
-                    icon: lastArg.item.img,
+                    icon: workflow.item.img,
                     origin: actor.uuid,
                     duration: {startTime: game.time.worldTime, seconds: 60},
                     changes: [