+++ scripts/creatures/life-drain.js
@@ -1,52 +1,50 @@
 /*
-	 The target must succeed on a DC 10 Constitution saving throw or its hit point maximum is reduced by an amount equal
+	 The target must succeed on a Constitution saving throw or its hit point maximum is reduced by an amount equal
 	 to the damage taken. This reduction lasts until the creature finishes a long rest. The target dies if this effect
 	 reduces its hit point maximum to 0.
 */
-const version = "10.0";
+const version = "11.0";
 const optionName = "Life Drain";
 const hpMaxFlag = "original-max-hp";
 
 try {
-    const lastArg = args[args.length - 1];
-
-    if (args[0] === "on") {
-        let workflow = MidiQOL.Workflow.getWorkflow(lastArg.origin);
-
-        if (actor) {
+    if (args[0].macroPass === "postActiveEffects") {
+        let target = workflow.failedSaves.first();
+        if (target) {
             const nectroticDamage = workflow.damageDetail.find(i => i.type === "necrotic");
 
             if (nectroticDamage.damage > 0) {
                 // apply hp max reduction
-                let currentMax = actor.system.attributes.hp.max;
+                let currentMax = target.actor.system.attributes.hp.max;
                 const newMax = Math.max(0, currentMax - nectroticDamage.damage);
 
                 // save away the original max hp
-                let existingFlag = actor.getFlag('world', hpMaxFlag);
+                let existingFlag = target.actor.getFlag('fvtt-trazzm-homebrew-5e', hpMaxFlag);
                 if (!existingFlag) {
-                    actor.setFlag('world', hpMaxFlag, currentMax);
+                    await target.actor.setFlag('fvtt-trazzm-homebrew-5e', hpMaxFlag, currentMax);
                 }
 
                 // reduce the hp max and check for death
-                await actor.update({"system.attributes.hp.max": newMax });
+                await target.actor.update({"system.attributes.hp.max": newMax });
                 ChatMessage.create({
-                    content: `${actor.name} is drained ${nectroticDamage.damage} Max HP`,
+                    content: `${target.name} is drained ${nectroticDamage.damage} Max HP`,
                     speaker: ChatMessage.getSpeaker({ actor: actor })});
 
                 // check for death
                 if (newMax === 0) {
-                    await actor.update({"system.attributes.hp.value": 0});
-                    await applyDeathEffect(lastArg.origin, actor);
+                    await target.actor.update({"system.attributes.hp.value": 0});
+                    await applyDeathEffect(workflow.origin, target.actor);
                 }
             }
+
         }
     }
     else if (args[0] === "off") {
-        const oldHpMax = actor.getFlag('world', hpMaxFlag);
+        const oldHpMax = actor.getFlag('fvtt-trazzm-homebrew-5e', hpMaxFlag);
         if (oldHpMax) {
+            await actor.unsetFlag('fvtt-trazzm-homebrew-5e', hpMaxFlag);
             await actor.update({"system.attributes.hp.max": oldHpMax });
             await ChatMessage.create({content: `${actor.name} Max HP returns to normal`});
-            await actor.unsetFlag('world', hpMaxFlag);
         }
     }