+++ scripts/spells/spell-spike-growth-support.js
@@ -1,6 +1,3 @@
-
-
-
 /*
 	When Entered
 */